devmapper: Remove waitRemove() function
So far we did not use UdevWait() function after device removal and we
instead would sit in a tight loop doing GetInfo and see if device actually
has been removed or not.

Now vbatts has inroduced use of UdevWait() in device removal path so 
there should not be any need to poll for device removal. Removing
waitRemove().

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>