devmapper: Do not try to remove a device if open count is non zero
Currently when we unmount a device we check for open count of device.
If it is not zero, we wait in a loop for some time for it to become
zero (To be precise, we try 1000 times at an interval of 10 ms). If it
still does not drop to zero, we log a warning message and continue to call
into libdm for device removal.

But the thing is that libdm also fails and returns with EBUSY (as device's
open count is not zero). We try 1000 times in a loop and then give up.

What's the point of trying 1000 time deletion if open count does not
drop to 0 after waiting for 10 seconds.

Do not call into libdm for device removal if open count of device does
not drop to zero. Simply return an error.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>