devmapper: Save device set metadata early during shutdown
When docker is being shutdown, save deviceset metadata first before
trying to remove the devices. Generally caller gives only 10 seconds
for shutdown to complete and then kills it after that. So if some device
is busy, we will wait 20 seconds for it removal and never be able to save
metadata. So first save metadata and then deal with device removal.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>