devmapper: Use a worker thread to cleanup deferred deleted devices
Right now we cleanup deferred deleted devices during startup. But this might
not be sufficient as one might not start docker for a long time and accumulate
lot of deferred deleted devices. This means lot of space in the pool will
unnecessarly be remain used.

Instead implemnet a worker thread which will try to cleanup deferred deleted
devices every 30 seconds (if there are deleted devices).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>