Make xfs as default container root filesystem
Currently ext4 is default root file system for docker. We try to override
it with xfs using docker-storage-setup but that works only when
docker-storage-setup is succesful out of the box. Many of the images
don't have volume groups or don't have free space in volume group so
there is no space to setup lvm thin pool hence docker-storage-setup fails
and hence ext4 is used by default.

This has couple of problems.

- We are not consistent. Sometime we use ext4 as container rootfs and
  other times we use xfs as rootfs.

- Currently docker ext4 is taking long time to do mkfs. Reason being that
  it is initializing all the metadata at mkfs time and that slows it down.
  Especially with 100G thin device on AWS storage, it takes very long time
  and systemd times out. In my experiment on AWS mkfs even on 10G device
  was taking more than 3 minutes.

So change default fs to xfs. For upstream discussion about it, refer to.

https://github.com/docker/docker/issues/17464
https://github.com/docker/docker/pull/17627

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>