devmapper: Get rid of ext4 lazy initialization
Lazy initialization of metadata writes lot of data during mkfs.ext4. For a
10G size of thin volume, it wrote around 300MB of data and took around 3
minutes with average write speed of 1.5MB/s. (This is thin pool on top of
loop devices in AWS). Now thin device size is 100G and that can take a
very long time forcing systemd to fail.

So get rid of preallocating metadata at mkfs time and fall back to the
default of lazy initialization. One side affect of this seems to be that
when we take snapshots of base image and container is running, then metadata
will be initialized and as many containers can share same image, there
will be multiple copies of metadata leading to wasted storage space.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>