An implementation of hard link based graph driver
This is a simple implementation of hardlinks graph driver. A lot of code
has simply been taken from overlay graph driver. This is meant to be
used when rootfs is supposed to read only. So always make sure to 
pass in --read-only option in docker while running a container if using
hardlinks graph driver. Otherwise it will corrupt underlying images.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>